+++ src/config/canvas_option.enabled.ml
@@ -1,0 +1,2 @@
+let detected = "yes"
+let enabled = true